IntelligentCustomerService/shop/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tortoise-orm==0.20.0
pydantic-settings==2.0.3
python-multipart==0.0.6
aiosqlite==0.19.0
